Innitial upload
The most notable change from previous parts is the move to a 5/16-18
hex head lead screw from the previous 1/4-10 hex head lag bolt. An
EZ-Lock threaded insert installs into the plunger mass which is also
longer than previous to increase moving mass and thus pitching moment.
Other modifications were made to the plunger mass to accommodate the
larger screw.
The bottom end of the plunger is now chamfered at 45deg to eliminate
the need for support material.
The plunger end has also been modified, it is a tighter fit with
anti-rotation spikes to eliminate the need to glue the rubber plunger.

The shaft coupling has been modified to accept the larger head of the
5/16 bolt. It also has a keyed back face. The servo horn fits perfectly
into the keyway reducing the likelihood of improper installation. The
screw holes and countersinks have also been tweaked. The holes
expanded, the countersinks deepened and chamfered for printability.

The rudder mount has been shrunken and a pad has been added to the
bottom face. this aids with printability and reduces print time.